Address host metrics review feedback
